튜토리얼, 음성인식 추가

+++ app/src/main/java/com/example/Touch/TouchPoint.java
@@ -1,11 +1,12 @@
 package com.example.Touch;
 
+//터치 좌표들을 한시적으로 get, set 하는 지점
 public class TouchPoint {
     private float x;
     private float y;
-    private int delay;
+    private long delay;
 
-    public TouchPoint(float x, float y, int delay) {
+    public TouchPoint(float x, float y, long delay) {
         this.x = x;
         this.y = y;
         this.delay = delay;
@@ -19,5 +20,5 @@ public class TouchPoint {
         return y;
     }
 
-    public int getDelay() { return delay; }
+    public long getDelay() { return delay; }
 }
